+++ imms/toxccs/__init__.py
@@ -1,0 +1,11 @@
+"""
+toxccs/__init__.py
+
+Ryan Nguyen (ryan97@uw.edu)
+
+description:
+        Python package with code for extracting and processing Waters .raw LC-IM-MS/MS data. 
+"""
+
+from .extract import RawProcessor
+from .multi_extract import MultiRawProcessor
