+++ imms/toxccs/extract/__init__.py
@@ -1,0 +1,9 @@
+"""
+toxccs/extract/__init__.py
+
+Ryan Nguyen (ryan97@uw.edu)
+
+Sub-package with utilities for extracting Waters .raw LC-IM-MS/MS data.
+"""
+
+from .RawProcessor import RawProcessor
